changed all ids to be longs instead of ints

+++ Models/IMochaDTOs/IMochaCandidateTestReport.cs
@@ -2,8 +2,8 @@ namespace IPR_BE.Models;
 
 public class CandidateTestReport
 {
-    public int testInvitationId { get; set; }
-    public int testId { get; set; }
+    public long testInvitationId { get; set; }
+    public long testId { get; set; }
     public string candidateEmail { get; set; } = "";
     public string candidateName { get; set; } = "";
     public DateTime attemptedOn { get; set; }
