+++ Models/IMochaDTOs/IMochaSection.cs
@@ -5,7 +5,7 @@ public class Section
         public double candidateScore { get; set; }
         public double sectionScore { get; set; }
         public int noOfQue { get; set; }
-        public int sectionID { get; set; }
+        public long sectionID { get; set; }
         public string sectionName { get; set; } = "";
         public int sectionTime { get; set; }
         public int sectionTimeTaken { get; set; }
